% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpTools.R
\name{clipbrd.read}
\alias{clipbrd.read}
\title{Copies from clipboard}
\usage{
clipbrd.read(sep = "\\t", quote = "\\"", stringsAsFactors = F)
}
\arguments{
\item{sep}{Column separator. Defaulted to \t.}
}
\value{
Gets clipboard.
}
\description{
Copies a tsv table from clipboard.
}
\examples{
clipbrd.read()
ds <- clipbrd.read(quote = "'")
}
