% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpTools.R
\name{clipbrd.write}
\alias{clipbrd.write}
\title{Copies to clipboard}
\usage{
clipbrd.write(x, sep = "\\t", quote = T)
}
\arguments{
\item{x}{tbl to copy.}

\item{sep}{Column separator. Defaulted to \t.}
}
\value{
Copies the object to the clipboard.
}
\description{
Copies an object to clipboard as a tsv table.
}
\examples{
clipbrd.write(ds)
ds \%>\%
  clipbrd.write()
}
