% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_clipbrd.R
\name{ClipbrdRead}
\alias{ClipbrdRead}
\title{Copies from clipboard}
\usage{
ClipbrdRead(sep = "\\t", quote = "\\"", stringsAsFactors = F, ...)
}
\arguments{
\item{sep}{Column separator. Defaulted to \t.}
}
\value{
Gets clipboard.
}
\description{
Copies a tsv table from clipboard.
}
\examples{
ClipbrdRead()
ds <- ClipbrdRead(quote = "'")
}
