% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_clipbrd.R
\name{ClipbrdWrite}
\alias{ClipbrdWrite}
\title{Copies to clipboard}
\usage{
ClipbrdWrite(x, sep = "\\t", quote = T, na = "", row.names = F,
  col.names = T, ...)
}
\arguments{
\item{x}{tbl to copy.}

\item{sep}{Column separator. Defaulted to \t.}
}
\value{
Copies the object to the clipboard.
}
\description{
Copies an object to clipboard as a tsv table.
}
\examples{
ClipbrdWrite(ds)
ds \%>\%
  ClipbrdWrite()
}
