% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_compare.R
\name{Compare}
\alias{Compare}
\title{Compare two dataframe for differences}
\usage{
Compare(x, y, inXbutNotY = T)
}
\arguments{
\item{x}{First data frame to compare.}

\item{y}{Second data frame to compare with.}

\item{inXbutNotY}{If True (default) informs to find differences in X and not in Y. If False infors to find differences in Y and not in X.}
}
\value{
Returns a tibble with differences with three new columns a. duplicates, count of matches based on a key. b. lhs_matches, count of matches in x for the entire row, c. rhs_matches, count of matches in y for the entire row.
}
\description{
Compare two data frames.
Both the data frames must have the same set of columns.
}
\examples{
Compare(x, y)
Compare(y, x)
}
