% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntpTools.R
\name{compare_table}
\alias{compare_table}
\title{Compare two dataframe for difference}
\usage{
compare_table(x, y, left_exclude = T)
}
\arguments{
\item{x}{A tbl to compare.}

\item{y}{Another tbl to compare with.}

\item{left_exclude}{Informs if left exclude comparison has to be used. If F, then uses right exclude join and is same as compare_table(y, x).}
}
\value{
Returns a tibble with the difference.
}
\description{
Compare two data frames for left exclude or right exclude comparison.
Both the data frames must have the same set of columns.
}
\examples{
compare_table(x, y)
compare_table(y, x)
}
