% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_compare.R
\name{CompareAndShowAll}
\alias{CompareAndShowAll}
\title{Compare two dataframe to show differences highlighted}
\usage{
CompareAndShowAll(x, y, col.names = c("lhs_matches", "rhs_matches"),
  check_duplicates_of = 1)
}
\arguments{
\item{x}{First data frame to compare.}

\item{y}{Second data frame to compare with.}
}
\value{
Returns a tibble with the difference.
}
\description{
Compares two data frames to show two additional columns x_match, y_match to inform the number of rows that matches against x and y. These columns informs if the row is present in x or y or both. Both the data frames must have the same set of columns.
}
