% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_compare.R
\name{CompareOraDataset}
\alias{CompareOraDataset}
\title{Compares two SQL resultsets from two database environments}
\usage{
CompareOraDataset(query, env1, env2, inXbutNotY = T)
}
\arguments{
\item{query}{SQL Query.}

\item{env1}{Environment string to the first database.}

\item{env2}{Environment string to the second database.}

\item{inXbutNotY}{If True (default) informs to find differences in X and not in Y. If False infors to find differences in Y and not in X.}
}
\value{
Returns a tibble of differences
}
\description{
Takes a SQL statement as an input and executes the SQL in two database environments, then compares the SQL resultset for differences.
}
