% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipbrd.R
\name{CopyFromClipboard}
\alias{CopyFromClipboard}
\title{Copies from clipboard}
\usage{
CopyFromClipboard(sep = "\\t", quote = "\\"", stringsAsFactors = F,
  header = T, ...)
}
\arguments{
\item{sep}{Column separator. Defaulted to tab.}

\item{quote}{Informs the quote character.}

\item{stringsAsFactors}{Informs if string has to be treated as factors.}

\item{header}{Informs if a header is the first row.}

\item{...}{Other parameters.}
}
\value{
Gets clipboard.
}
\description{
Copies a tsv table from clipboard.
}
\examples{
CopyFromClipboard()
ds <- CopyFromClipboard(quote = "'")
}
