% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crime}
\alias{crime}
\title{Table: FBI crime data}
\format{A data frame with 20 years of crime from 1994 through 2013.
Columns are: Year, Population, Violent_crime, Violent_crime_rate, Murder_and_nonnegligent_manslaughter, Murder_and_nonnegligent_manslaughter_rate, Rape_legacy_definition, Rape_legacy_definition_rate, Robbery, Robbery_rate, Aggravated_assault, Aggravated_assault_rate, Property_crime, Property_crime_rate, Burglary, Burglary_rate, Larceny_theft, Larceny_theft_rate, Motor_vehicle_theft, Motor_vehicle_theft_rate}
\source{
FBI
}
\usage{
crime
}
\description{
A dataset containing crime data
\itemize{
  \item Year.
  \item Population.
  ...
}
}
\keyword{datasets}
