% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{employment_population_ratio}
\alias{employment_population_ratio}
\title{Table: Employment Population Ratio}
\format{A data frame with employment ratio from 2007 through 2016.}
\source{
Department of Labor
}
\usage{
employment_population_ratio
}
\description{
A dataset containing ratio between population and employment
\itemize{
  \item Year.
  \item Months of the Year.
}
}
\keyword{datasets}
