% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_ntp.R
\name{env.compare}
\alias{env.compare}
\title{Compares two SQL resultsets from two database environments}
\usage{
env.compare(query, connection1, connection2, left_exclude = T)
}
\arguments{
\item{query}{SQL Query.}

\item{connection1}{Connection string to the first database.}

\item{connection2}{Connection string to the second database.}

\item{left_exclude}{left_exclude Informs if left exclude comparison has to be used. If F, then uses right exclude join and is same as table.compare(y, x).}
}
\value{
Returns a tibble of differences
}
\description{
Takes a SQL statement as an input and executes the SQL in two database environments, then compares the SQL resultset for differences.
}
