% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{Freq}
\alias{Freq}
\title{Frequency table of all columns}
\usage{
Freq(x, records = 10, as = "l")
}
\arguments{
\item{x}{Dataset to be processed.}

\item{records}{Number of rows to show in the resulting table}

\item{as}{return as "l" list or "d" dataset}
}
\value{
Returns a frequency table in a key/value pair format for all the columns. Number of columns returned is length of the dataset x 2.
}
\description{
Generates a frequency distribution table of all columns in a data set.
}
\examples{
Freq(census)
Freq(census)$State
Freq(census, 50, "d")
Freq(census, 100, "d") #will throw an error, use Freq(census, 100, "l") instead

}
