% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_freqsummary.R
\name{FreqTable}
\alias{FreqTable}
\title{Frequency table of all columns}
\usage{
FreqTable(x, records = 10)
}
\arguments{
\item{x}{Dataset to be processed.}

\item{records}{Number of rows to show in the resulting table}
}
\value{
Returns a frequency table in a key/value pair format for all the columns. Number of columns returned is length of the dataset x 2.
}
\description{
Generates a frequency distribution table of all columns in a data set.
}
