% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geompointedline.R
\name{geom_pointedline}
\alias{geom_pointedline}
\title{Creates a pointed lined on ggplot}
\usage{
geom_pointedline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show.legend = NA, inherit.aes = TRUE,
  size = 1.5, ...)
}
\examples{
ggplot() +
 geom_pointedline(
   data = filter(LifeExpectancy, Country.Code == "WLD") \%>\% drop_na(),
   aes(x = Year, y = Expectancy, color = Country.Code)
 )  +
   geom_pointedline(
     data = filter(LifeExpectancy, Country.Code == "USA")  \%>\% drop_na(),
     aes(x = Year, y = Expectancy, color = Country.Code)
   )
}
