% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_xml.R
\name{get_all_node_text}
\alias{get_all_node_text}
\title{Gets all XML node text}
\usage{
get_all_node_text(x)
}
\arguments{
\item{x}{XML document.}
}
\value{
Returns xpath to a node, node name and text value in a tibble.
}
\description{
Parses through all XML end nodes and returns the text of those nodes in a tibble.
}
\examples{
get_all_node_text(xml2::read_xml("https://www.w3schools.com/xml/simple.xml")) \%>\%
  View()
}
