% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_compare.R
\name{\%=\%}
\alias{\%=\%}
\title{Compares two datasets}
\usage{
x %=% y
}
\arguments{
\item{x}{Dataset to compare.}

\item{y}{Dataset to compare with,}
}
\value{
Returns a dataset after merging and two column that denotes which of the merged rows exists in x or y or both.
}
\description{
Merges and compares two datasets.
}
