% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{income}
\alias{income}
\title{Table: Income Dataset}
\format{A data frame with income data from 1967 through 2015.}
\source{
Census Bureau
}
\usage{
income
}
\description{
A dataset containing income slabs
\itemize{
  \item Year.
  \item Households.
  ...
}
}
\keyword{datasets}
