% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{labs_cuckoo}
\alias{labs_cuckoo}
\title{Label a ggplot}
\usage{
labs_cuckoo(xlab = "value", ylab = "freq", title = "title",
  subtitle = paste("Dated", Sys.Date()), source = "Unknown",
  dated = Sys.Date())
}
\arguments{
\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{source}{Label for source that will appear as a footnote.}

\item{dated}{Label for date that will appear as a footnote.}
}
\value{
Returns the label
}
\description{
A labeler to add title, subtitle, axis titles and footnote to a ggplot.
}
\examples{
census \%>\%
  ggplot(aes(x = State, fill = State)) +
  geom_bar(stat = "count") +
  coord_flip() +
  labs_cuckoo(title = "County Frequency by State", subtitle = NULL, xlab = "State") +
  theme_cuckoo(font_family = "Tahoma",legend_position = "none")
}
