% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LifeExpectancy}
\alias{LifeExpectancy}
\title{Table: Life Expectancy}
\format{A data frame with life expectacny from 1960 through 2015.}
\source{
World Bank @ http://data.worldbank.org/indicator/SP.DYN.LE00.IN
}
\usage{
LifeExpectancy
}
\description{
A dataset containing Life Expectancy by Country or regions (like world, Arab world, etc) between 1960 and 2015
\itemize{
  \item Country.
  \item Year.
  \item Life Expectancy (Age).
}
}
\keyword{datasets}
