% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{ODBCRun}
\alias{ODBCRun}
\title{Runs a SQL query using ODBC driver}
\usage{
ODBCRun(query, dsn = "mydsn")
}
\arguments{
\item{query}{SQL Query to execute.}

\item{dsn}{Data source name.}
}
\value{
SQL resultset in a dataframe
}
\description{
Runs a SQL query in a database using the ODBC driver and returns a resultset. Before hand a DSN has to be created
under Control Panel -> Administrative tools.
}
\examples{
ODBCRun("SELECT * FROM ALL_TABLES", "myDSN")
}
