% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oraConnect.R
\name{ora.from_date}
\alias{ora.from_date}
\title{Converts an Oracle date (dd-MMM-yy) into a Date format.}
\usage{
ora.from_date(x, century = F)
}
\arguments{
\item{x}{Date string to be converted.}

\item{century}{Informs if the string has the century included (dd-MM-yyyy). Default is F.}
}
\value{
Returns the date string as a date
}
\description{
Converts an Oracle date (dd-MMM-yy) into a Date format.
}
\examples{
ora.from_date("01-JAN-16")
ora.from_date("01-JAN-2017", T)
}
