% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oraConnect.R
\name{ora.run}
\alias{ora.run}
\title{Runs a SQL query in preferred environment.}
\usage{
ora.run(query, env)
}
\arguments{
\item{query}{SQL Query to execute.}

\item{env}{Environment name. Defaults to dev. Other values are stg, tst, prod.}
}
\value{
SQL resultset as a tibble
}
\description{
Runs a SQL query in Oracle and returns a resultset.
}
\examples{
ora.run("SELECT * FROM ALL_TABLES", "stg")

}
