% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oraConnect.R
\name{ora.run_query}
\alias{ora.run_query}
\title{Runs a SQL query in Oracle}
\usage{
ora.run_query(query, host_name, port = "1521", sid = "xe", user_name, pwd)
}
\arguments{
\item{query}{SQL Query to execute.}

\item{host_name}{Host name.}

\item{port}{Port name. Defaults to port 1521.}

\item{sid}{Service Id. Defaults to xe.}

\item{user_name}{User name for the connection in plain text.}

\item{pwd}{Password for the connection in plain text.}
}
\value{
SQL resultset in a tibble
}
\description{
Runs a SQL query in Oracle and returns a resultset.
}
\examples{
ora.run_query("SELECT * FROM ALL_TABLES", "act", 1580, xe, "scott", "tiger")

}
