% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{OraTableDesc}
\alias{OraTableDesc}
\title{Gets table description from Oracle db}
\usage{
OraTableDesc(db_table, env = "dev")
}
\arguments{
\item{db_table}{Name of the database table.}

\item{env}{Environment string to use for forming the connection string.}
}
\value{
Returns a data frame
}
\description{
Table description of an Oracle table is retrieved in a concise format.
}
\examples{
OraTableDesc("EMP", "dev")
}
