% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon.R
\name{Recon}
\alias{Recon}
\title{Compare two dataframe to show differences highlighted}
\usage{
Recon(x, y, col.names = c("lhs_matches", "rhs_matches"),
  check_duplicates_of = 1)
}
\arguments{
\item{x}{First data frame to compare.}

\item{y}{Second data frame to compare with.}

\item{col.names}{Informs the title of the new columns informs if matches exists. Defaulted to lhs_matches and rhs_matches.}

\item{check_duplicates_of}{Informs the primary key columnindex for which duplicates has to be checked. The result will be listed in a column named duplicates.}
}
\value{
Returns a tibble with the difference.
}
\description{
Compares two data frames to show two additional columns lhs_matches, rhs_matches to inform the number of rows that matches against x and y. These columns informs if the row is present in x or y or both. Both the data frames must have the same set of columns.
}
\examples{
Recon(list1, list2)
}
