% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{Stat}
\alias{Stat}
\title{Shows a quick summary of a data frame}
\usage{
Stat(df, sample = 20, sample_delim = "; ")
}
\arguments{
\item{df}{Data frame.}

\item{sample}{Number of samples from each column. Defaults to 20.}

\item{sample_delim}{Delimiter for each sample.}
}
\value{
Stat(census)
}
\description{
Shows element type, row count, na row count, unique row count, mean, min, Q1, median, Q3, max, min string length, max string length and set of samples from each of the columns in a data frame.
}
