% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{StringToDate}
\alias{StringToDate}
\title{Converts an date string to a date}
\usage{
StringToDate(x, century = F)
}
\arguments{
\item{x}{Date string to be converted.}

\item{century}{Informs if the string has the century included (dd-MM-yyyy). Default is F.}
}
\value{
Returns the date string as a date
}
\description{
Converts a string in date format(dd-MMM-yy) into a R Date.
}
\examples{
StringToDate("01-JAN-16")
StringToDate("01-JAN-2017", T)
}
