% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_ntp.R
\name{table.compare}
\alias{table.compare}
\title{Compare two dataframe for difference}
\usage{
table.compare(x, y, left_exclude = T)
}
\arguments{
\item{x}{A tbl to compare.}

\item{y}{Another tbl to compare with.}

\item{left_exclude}{Informs if left exclude comparison has to be used. If F, then uses right exclude join and is same as table.compare(y, x).}
}
\value{
Returns a tibble with the difference.
}
\description{
Compare two data frames for left exclude or right exclude comparison.
Both the data frames must have the same set of columns.
}
\examples{
table.compare(x, y)
table.compare(y, x)
}
