% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{text_cuckoo}
\alias{text_cuckoo}
\title{Add text in the plot}
\usage{
text_cuckoo(lbl, font_family = "Tahoma", hjust = 1, ...)
}
\arguments{
\item{lbl}{Column value to be displayed as label.}

\item{font_family}{Font family. Defaulted to Tahoma.}

\item{hjust}{Horizontal justification as Left Justified.}
}
\value{
Returns the text
}
\description{
Adds text to the histograms or points.
}
\examples{
census \%>\%
  ggplot(aes(x = State, color = State)) +
  geom_point(stat = "count", size = 8) +
  labs_cuckoo(title = "County Frequency by State", subtitle = NULL, xlab = "State") +
  text_cuckoo("..count..", color = "black", hjust = 0.5) +
  theme_cuckoo(font_family = "Tahoma",legend_position = "none", axis.text.x = element_text(angle = 45, hjust = 1))
}
