% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_cuckoo}
\alias{theme_cuckoo}
\title{Applies the ggplots theme "cuckoo"}
\usage{
theme_cuckoo(font_family = "Arial", base_size = 10,
  legend_position = c(0.01, 0.99), ...)
}
\arguments{
\item{font_family}{Informs the font family to be used as a base. Defaulted to Tahoma.}

\item{base_size}{Informs the base font size. Defaulted to size 10.}

\item{legend_position}{Informs the postion of the legend. Use "none" to hide the legend.}
}
\value{
Returns the theme
}
\description{
A theme that can be applied to a ggplot.
}
\examples{
census \%>\%
  ggplot(aes(x = State, fill = State)) +
  geom_bar(stat = "count") +
  coord_flip() +
  theme_cuckoo(font_family = "Tahoma",legend_position = "none")
}
