% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{unemployment_rate}
\alias{unemployment_rate}
\title{Table: Unemployment Rate}
\format{A data frame with unemployment rate from 2007 through 2016.}
\source{
Department of Labor
}
\usage{
unemployment_rate
}
\description{
A dataset containing unemployment rate
\itemize{
  \item Year.
  \item Months of the Year.
}
}
\keyword{datasets}
