% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{XmlGetValues}
\alias{XmlGetValues}
\title{Gets all XML node text}
\usage{
XmlGetValues(x, nodepath = "*")
}
\arguments{
\item{x}{XML document.}

\item{nodepath}{Xpath to node for which text has to be collected. Defaults to * that will collect text of all nodes.}
}
\value{
Returns xpath to a node, node name and text value in a tibble.
}
\description{
Parses through all XML end nodes and returns the text of those nodes in a tibble.
}
\examples{
XmlGetValues(xml2::read_xml("https://www.w3schools.com/xml/simple.xml")) \%>\%
  View()
XmlGetValues(xml2::read_xml("https://www.w3schools.com/xml/simple.xml"), "calories") \%>\%
  View()
}
