% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipbrd.R
\name{CopyToClipboard}
\alias{CopyToClipboard}
\title{Copies to clipboard}
\usage{
CopyToClipboard(x, sep = "\\t", quote = T, na = "", row.names = F,
  col.names = T, ...)
}
\arguments{
\item{x}{tbl to copy.}

\item{sep}{Column separator. Defaulted to tab.}

\item{quote}{T (default) informs that the result will be within quotes.}

\item{na}{Value for NAs. Default is an empty string.}

\item{row.names}{Informs if row names must be included.}

\item{col.names}{Informs if column names must be included.}

\item{...}{Other parameters.}
}
\value{
Copies the object to the clipboard.
}
\description{
Copies an object to clipboard as a tsv table.
}
\examples{
CopyToClipboard(census)
census \%>\%
  CopyToClipboard()
}
