% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{labs_cuckoo}
\alias{labs_cuckoo}
\title{Label a ggplot}
\usage{
labs_cuckoo(source = "Unknown", dated = Sys.Date(), ...)
}
\arguments{
\item{source}{Label for source that will appear as a footnote.}

\item{dated}{Label for date that will appear as a footnote.}
}
\value{
Returns the label
}
\description{
A labeler to add title, subtitle, axis titles and footnote to a ggplot.
}
\examples{
census \%>\%
  ggplot(aes(x = State, fill = State)) +
  geom_bar(stat = "count") +
  coord_flip() +
  labs_cuckoo(label = "County Frequency by State", subtitle = NULL, x = "State", y = "Frequency") +
  theme_cuckoo(font_family = "Tahoma",legend_position = "none")
}
